+++ Tests/Runtime/LSL/LSLStreamResolverTests.cs
@@ -1,4 +1,4 @@
-using System;
+using System.Collections;
 using UnityEngine;
 using UnityEngine.TestTools;
 using LSL;
@@ -6,39 +6,18 @@ using BCIEssentials.Tests.Utilities;
 using NUnit.Framework;
 
 using static BCIEssentials.LSLFramework.LSLStreamResolver;
-using static System.Diagnostics.Process;
-using System.Collections;
 
 namespace BCIEssentials.Tests
 {
-    public class LSLStreamResolverTests: PlayModeTestRunnerBase
+    public class LSLStreamResolverTests: LSLOutletTestRunner
     {
-        private StreamOutlet _testOutlet;
-        private const string OutletName = "Unity Testing Outlet";
-        private const string OutletType = "Test Markers";
-
-
-        [UnitySetUp]
-        public override IEnumerator TestSetup()
-        {
-            yield return base.TestSetup();
-            _testOutlet = BuildOutlet(OutletName, OutletType);
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            _testOutlet.Dispose();
-        }
-
-
         [Test]
         public void TryResolveByType_WhenOutletAvailable_ThenReturnsTrue()
-        => Assert.IsTrue(TryResolveByType(OutletType, out _));
+        => Assert.IsTrue(TryResolveByType(TestOutletType, out _));
 
         [Test]
         public void TryResolveByName_WhenOutletAvailable_ThenReturnsTrue()
-        => Assert.IsTrue(TryResolveByName(OutletName, out _));
+        => Assert.IsTrue(TryResolveByName(TestOutletName, out _));
 
 
         [Test]
@@ -71,24 +50,6 @@ namespace BCIEssentials.Tests
 
             Assert.IsNotNull(resolvedStreamInfo);
             outlet.Dispose();
-        }
-
-
-        private StreamOutlet BuildOutlet
-        (
-            string streamName = "Unity Testing Outlet",
-            string streamType = "Test Markers"
-        )
-        {
-            string deviceID = SystemInfo.deviceUniqueIdentifier;
-            int processID = GetCurrentProcess().Id;
-            var streamInfo = new StreamInfo
-            (
-                streamName, streamType,
-                channel_format: channel_format_t.cf_string,
-                source_id: $"{deviceID}-Unity-{processID}"
-            );
-            return new StreamOutlet(streamInfo);
         }
     }
 }
