--- Tests/Runtime/PlayModeTestRunnerBase.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: e0a11b0db6254180b8d9b8a3187eefed
-timeCreated: 1672171625
