+++ Tests/Utilities/LSLOutletTestRunner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5109424c5d50bd248b253700fa8a7ed9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
